+++ src/main/java/org/wedding/application/port/in/usecase/CreateCardUseCase.java
@@ -1,9 +1,9 @@
-package org.wedding.usecase;
+package org.wedding.application.port.in.usecase;
 
 import java.util.logging.Logger;
 
 import org.wedding.domain.CardContainerRepository;
-import org.wedding.domain.cardboard.CardBoard;
+import org.wedding.domain.CardBoard;
 
 public class CreateCardUseCase {
     private final CardContainerRepository cardContainerRepository;
