+++ src/main/java/org/wedding/domain/CardContainerRepository.java
@@ -1,7 +1,5 @@
 package org.wedding.domain;
 
-import org.wedding.domain.cardboard.CardBoard;
-
 public interface CardContainerRepository {
     CardBoard findByUserId(String userId);
 }
