+++ src/main/java/org/wedding/domain/CardFactory.java
@@ -1,4 +1,4 @@
-package org.wedding.domain.card;
+package org.wedding.domain;
 
 public class CardFactory {
