+++ src/main/java/org/wedding/domain/CardStatus.java
@@ -1,4 +1,4 @@
-package org.wedding.domain.card;
+package org.wedding.domain;
 
 public enum CardStatus {
     BACKLOG, PROGRESS, DONE
