+++ src/main/java/org/wedding/domain/Todo.java
@@ -1,4 +1,4 @@
-package org.wedding.domain.user;
+package org.wedding.domain;
 
 import lombok.Getter;
