+++ src/test/java/org/wedding/domain/CardBoardTest.java
@@ -6,9 +6,6 @@ import java.util.List;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.wedding.domain.card.Card;
-import org.wedding.domain.cardboard.CardBoard;
-import org.wedding.domain.card.CardStatus;
 
 class CardBoardTest {
