+++ src/test/java/org/wedding/domain/CardTest.java
@@ -5,7 +5,6 @@ import static org.assertj.core.api.Assertions.*;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.wedding.domain.card.Card;
 
 @Disabled
 class CardTest {
