+++ src/test/java/org/wedding/user/service/UserServiceTest.java
@@ -9,10 +9,9 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import org.wedding.domain.user.User;
+import org.wedding.domain.User;
 import org.wedding.user.repository.UserRepository;
 import org.wedding.user.repository.mybatis.dto.request.SignUpRequest;
-import org.wedding.user.service.UserService;
 
 @ExtendWith(MockitoExtension.class)
 class UserServiceTest {
